move page unit tests to page folder
