+++ test/unit/views/organisations/issueDetailsPage.test.js
@@ -1,10 +1,10 @@
 import { describe, it, expect } from 'vitest'
-import { setupNunjucks } from '../../src/serverSetup/nunjucks.js'
+import { setupNunjucks } from '../../../../src/serverSetup/nunjucks.js'
 import { JSDOM } from 'jsdom'
-import { runGenericPageTests } from './generic-page.js'
-import config from '../../config/index.js'
-import { OrgIssueDetails } from '../../src/routes/schemas.js'
-import mocker from '../utils/mocker.js'
+import { runGenericPageTests } from '../../generic-page.js'
+import config from '../../../../config/index.js'
+import { OrgIssueDetails } from '../../../../src/routes/schemas.js'
+import mocker from '../../../utils/mocker.js'
 
 const nunjucks = setupNunjucks({})
